requirements.txt
pandas
numpy
tensorflow-cpu
streamlit
folium
textblob
matplotlib
requests